Inline Case files for EqualsAlwaysReturnsTrueOrFalseSpec
